example code of new API (here as "C" only)
This commit contains a sample code that shows the new API.

For here now as "C", in "C++" the interface remains almost 100% as before.

The biggest addon differences are here with:
- The addon is run independently of Kodi
  - This ensures that Kodi is not destroyed in the event of a crash
  - In addition, in the event of an update, it is ensured that all its data is deleted and processes are stopped after the add-on is closed
  - It now allows other languages, as here only as "C", and it can be extended more easily for others, such as Java
- Only a number is used as a version (similar to Android) and should, if possible, allow downward compatibility with older add-ons.
  - Since processing between Kodi and addon is now again done via a special Lib, this should continue to make the APIs available for different versions
    - The Lib code should be created automatically as far as possible, so that what does not have to be edited by hand

This project on Kodi is very large and still needs a little time to be usable.
In addition, in the new way, it is also possible to use binary on other addon types (test on a new Kodi service addon as "C" already OK).

Runtime tests with here was OK.